+++ experimental/ragged_inference/ragged_inference/seq_kv_cache.py
@@ -7,8 +7,8 @@ from functools import lru_cache
 from typing import List, Tuple
 
 import torch
-from ragged_inference_v2.garbage_pad_ragged_acts import RaggedActivations
-from ragged_inference_v2.test_utils import assert_eq, bf16_cuda
+from ragged_inference.garbage_pad_ragged_acts import RaggedActivations
+from ragged_inference.test_utils import assert_eq, bf16_cuda
 
 # Whenever we need to grow the tensor, we extend by this amount extra so that we don't
 # have to do expensive memcopies as frequently as we would otherwise
